@@ -32,15 +32,15 @@
     "@fortawesome/fontawesome-free-regular": "^5.0.13",
     "@fortawesome/fontawesome-free-solid": "^5.0.6",
     "@ng-select/ng-select": "^2.3.6",
-    "apollo-angular": "^1.1.2",
-    "apollo-angular-link-http": "^1.1.1",
+    "apollo-angular": "^1.3.0",
+    "apollo-angular-link-http": "^1.2.0",
     "apollo-cache-inmemory": "^1.2.7",
-    "apollo-client": "^2.3.7",
+    "apollo-client": "^2.4.1",
     "apollo-link-ws": "^1.0.8",
     "core-js": "^2.5.4",
     "file-saver": "^1.3.8",
     "getstream": "^3.19.0",
-    "graphql": "^0.13.2",
+    "graphql": "^14.0.2",
     "graphql-tag": "^2.9.2",
     "rxjs": "^6.2.2",
     "subscriptions-transport-ws": "^0.9.14",
@@ -69,7 +69,7 @@
     "protractor": "~5.3.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.7.2",
+    "typescript": "^2.8.4",
     "webpack-bundle-analyzer": "^2.9.2"
   }
 }
