@@ -60,7 +60,7 @@ export class TeamClientInitializer extends AbstractApolloClient {
 		);
 
 		// combine tokens & uri
-		combineLatest(accessToken$, uri$)
+		zip(accessToken$, uri$)
 			.subscribe(([token, uri]) => super.initClient(uri, Client.TEAM, token));
 
 		// when no team selected we also destroy the client
